Move teleport's `ThemeProvider` to teleport
