+++ web/packages/design/src/utils/testing.tsx
@@ -30,13 +30,15 @@ import {
 import userEvent from '@testing-library/user-event';
 import { MemoryRouter as Router } from 'react-router-dom';
 
-import ThemeProvider from 'design/ThemeProvider';
+import { StaticThemeProvider } from 'design/ThemeProvider';
 import { darkTheme } from 'design/theme';
 import '@testing-library/jest-dom';
 import 'jest-styled-components';
 
 function Providers({ children }: { children: React.ReactElement }) {
-  return <ThemeProvider theme={darkTheme}>{children}</ThemeProvider>;
+  return (
+    <StaticThemeProvider theme={darkTheme}>{children}</StaticThemeProvider>
+  );
 }
 
 function render(ui: React.ReactElement<any>, options?: RenderOptions) {
