+++ web/packages/teleport/src/Console/DocumentKubeExec/DocumentKubeExec.test.tsx
@@ -16,15 +16,13 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 import React from 'react';
-import { render, screen } from '@testing-library/react';
+import { screen } from '@testing-library/react';
 
 import '@testing-library/jest-dom';
 
-import ThemeProvider from 'design/ThemeProvider';
 import 'jest-canvas-mock';
-import { darkTheme } from 'design/theme';
 
-import { act } from 'design/utils/testing';
+import { act, render } from 'design/utils/testing';
 
 import { ContextProvider } from 'teleport';
 import { TestLayout } from 'teleport/Console/Console.story';
@@ -65,13 +63,11 @@ describe('DocumentKubeExec', () => {
     const { ctx, consoleCtx } = getContexts();
 
     render(
-      <ThemeProvider theme={darkTheme}>
-        <ContextProvider ctx={ctx}>
-          <TestLayout ctx={consoleCtx}>
-            <DocumentKubeExec doc={baseDoc} visible={true} />
-          </TestLayout>
-        </ContextProvider>
-      </ThemeProvider>
+      <ContextProvider ctx={ctx}>
+        <TestLayout ctx={consoleCtx}>
+          <DocumentKubeExec doc={baseDoc} visible={true} />
+        </TestLayout>
+      </ContextProvider>
     );
   };
