+++ web/packages/teleport/src/services/storageService/storageService.ts
@@ -22,7 +22,7 @@ import { Theme } from 'gen-proto-ts/teleport/userpreferences/v1/theme_pb';
 
 import { OnboardUserPreferences } from 'gen-proto-ts/teleport/userpreferences/v1/onboard_pb';
 
-import { getPrefersDark } from 'design/ThemeProvider';
+import { getPrefersDark } from 'teleport/ThemeProvider';
 
 import { BearerToken } from 'teleport/services/websession';
 import { OnboardDiscover } from 'teleport/services/user';
